cloud_function/src/requirements.txt
google-cloud-storage
pandas
gcsfs
google-cloud-bigquery